limit scope; nits
