+++ packages/react-native-reanimated/jest/resolver.js
@@ -7,7 +7,6 @@ const WEB_ONLY_IN_JEST = new Set([
   'ConfigHelper',
   'UpdateLayoutAnimations',
   'useAnimatedRef',
-  'useAnimatedStyle',
 ]);
 
 /** @type {import('jest-resolve').SyncResolver} */
